Visual improvements Pick cousines in tabs Icons Place orders correctly Animations Ask for contact name and delivery address

+++ Hackaton/Features/Product/Model/ProductResponseModel.swift
@@ -14,8 +14,8 @@ struct ProductResponseModel: Codable, Equatable {
         return lhs.id == rhs.id
     }
     
-    let id: Int
-    let storeId: Int
+    let id: Int64
+    let storeId: Int64
     let name: String
     let description: String
     let price: Double
